Master sync works, and health is displayed and handled properly.
The issue with this was just too much client control. Some things might not work still, but they may just need to be toggled true via GameMaster.OnLevelWasLoaded
I went through and toggled EVERY script off with the exception of a new GameMaster script. This script checks if it's a server or client, and then toggles the needed scripts.
The existing code works fine. It was just the client and server all attempting to do the same things that was causing conflicts.